+++ examples/dataset/README.md
@@ -3,15 +3,8 @@
 This directory contains example Kolena Datasets integrations for different machine learning workflows.
 
 - [Binary and Multiclass Classification](./classification): open source classification models tested using the
-  [CIFAR-10](https://www.cs.toronto.edu/~kriz/cifar.html) and the
-  [Dogs vs. Cats](https://www.kaggle.com/c/dogs-vs-cats) datasets
-
+[Dogs vs. Cats](https://www.kaggle.com/c/dogs-vs-cats) and the
+[CIFAR-10](https://www.cs.toronto.edu/~kriz/cifar.html) datasets
 - [Question Answering](./question_answering): GPT models tested using the
-  [TruthfulQA (open-domain)](https://github.com/sylinrl/TruthfulQA) and the
-  [HaluEval (closed-domain)](https://github.com/RUCAIBox/HaluEval/tree/main/evaluation) datasets
-- [Text Summarization](./text_summarization): OpenAI's [GPT-3](https://platform.openai.com/docs/models/gpt-3) and
-  [GPT-3.5](https://platform.openai.com/docs/models/gpt-3-5) model families tested on the
-  [CNN-DailyMail](https://paperswithcode.com/dataset/cnn-daily-mail-1) dataset for text summarization workflow
-- [Rain Forecast](./rain_forecast): rain forecast models tested using the
-  [Rain in Australia](https://www.kaggle.com/datasets/jsphyg/weather-dataset-rattle-package)
-- [Speaker Diarization](./speaker_diarization/): speaker diarization using the [ICSI-Corpus](https://groups.inf.ed.ac.uk/ami/icsi/) dataset
+[TruthfulQA (open-domain)](https://github.com/sylinrl/TruthfulQA) and the
+[HaluEval (closed-domain)](https://github.com/RUCAIBox/HaluEval/tree/main/evaluation) datasets
